broke Orb Container and Orbs into their own components, changed Main and About pages to be in pages folder
