+++ client/index.js
@@ -7,5 +7,6 @@ import reactDom from 'react-dom';
 import App from './components/App.jsx';
 
 
-//mount our React app to the index.html at root element
+// mount our React app to the index.html at root element. 
+// The || is because React Testing Library doesn't load index.html and therefore can't find #root
 reactDom.render(<App/>, document.querySelector('#root') || document.createElement('div'));
